[mcache] Incorporate Feedback
