+++ mcache/tests/test_check.py
@@ -4,218 +4,12 @@
 
 import subprocess
 import os
-import time
 import pytest
-import bmemcached
-from bmemcached.exceptions import MemcachedException
 
 from datadog_checks.utils.platform import Platform
-from datadog_checks.mcache import Memcache
 
-from common import (HERE, PORT, SERVICE_CHECK, HOST, USERNAME, PASSWORD, DOCKER_SOCKET_DIR, DOCKER_SOCKET_PATH,
-                    UNIXSOCKET_DIR, UNIXSOCKET_PATH)
-
-GAUGES = [
-    "memcache.total_items",
-    "memcache.curr_items",
-    "memcache.limit_maxbytes",
-    "memcache.uptime",
-    "memcache.bytes",
-    "memcache.curr_connections",
-    "memcache.connection_structures",
-    "memcache.threads",
-    "memcache.pointer_size",
-    # Computed metrics
-    "memcache.get_hit_percent",
-    "memcache.fill_percent",
-    "memcache.avg_item_size"
-]
-
-RATES = [
-    "memcache.rusage_user_rate",
-    "memcache.rusage_system_rate",
-    "memcache.cmd_get_rate",
-    "memcache.cmd_set_rate",
-    "memcache.cmd_flush_rate",
-    "memcache.get_hits_rate",
-    "memcache.get_misses_rate",
-    "memcache.delete_misses_rate",
-    "memcache.delete_hits_rate",
-    "memcache.evictions_rate",
-    "memcache.bytes_read_rate",
-    "memcache.bytes_written_rate",
-    "memcache.cas_misses_rate",
-    "memcache.cas_hits_rate",
-    "memcache.cas_badval_rate",
-    "memcache.total_connections_rate",
-    "memcache.listen_disabled_num_rate"
-]
-
-# Not all rates/gauges reported by memcached test instance.
-# This is the subset available with the default config/version.
-ITEMS_RATES = [
-    "memcache.items.evicted_rate",
-    "memcache.items.evicted_nonzero_rate",
-    "memcache.items.expired_unfetched_rate",
-    "memcache.items.evicted_unfetched_rate",
-    "memcache.items.outofmemory_rate",
-    "memcache.items.tailrepairs_rate",
-    "memcache.items.reclaimed_rate",
-    "memcache.items.crawler_reclaimed_rate",
-    "memcache.items.lrutail_reflocked_rate",
-    "memcache.items.moves_to_warm_rate",
-    "memcache.items.moves_to_cold_rate",
-    "memcache.items.moves_within_lru_rate",
-    "memcache.items.direct_reclaims_rate",
-]
-
-ITEMS_GAUGES = [
-    "memcache.items.number",
-    "memcache.items.number_hot",
-    "memcache.items.number_warm",
-    "memcache.items.number_cold",
-    "memcache.items.age",
-    "memcache.items.evicted_time",
-]
-
-SLABS_RATES = [
-    "memcache.slabs.get_hits_rate",
-    "memcache.slabs.cmd_set_rate",
-    "memcache.slabs.delete_hits_rate",
-    "memcache.slabs.incr_hits_rate",
-    "memcache.slabs.decr_hits_rate",
-    "memcache.slabs.cas_hits_rate",
-    "memcache.slabs.cas_badval_rate",
-    "memcache.slabs.touch_hits_rate",
-    "memcache.slabs.used_chunks_rate",
-]
-
-SLABS_GAUGES = [
-    "memcache.slabs.chunk_size",
-    "memcache.slabs.chunks_per_page",
-    "memcache.slabs.total_pages",
-    "memcache.slabs.total_chunks",
-    "memcache.slabs.used_chunks",
-    "memcache.slabs.free_chunks",
-    "memcache.slabs.free_chunks_end",
-    "memcache.slabs.mem_requested",
-]
-
-SLABS_AGGREGATES = [
-    "memcache.slabs.active_slabs",
-    "memcache.slabs.total_malloced",
-]
-
-
-@pytest.fixture(scope="session")
-def memcached():
-    """
-    Start a standalone Memcached server.
-    """
-    env = os.environ
-    env['PWD'] = HERE
-    docker_compose_file = os.path.join(HERE, 'compose', 'docker-compose.yaml')
-
-    subprocess.check_call(["docker-compose", "-f", docker_compose_file, "up", "-d", "memcached"], env=env)
-    attempts = 0
-    while True:
-        if attempts > 10:
-            raise Exception("Memcached boot timed out!")
-
-        mc = bmemcached.Client(["{}:{}".format(HOST, PORT)], USERNAME, PASSWORD)
-        try:
-            mc.set("foo", "bar")
-        except MemcachedException:
-            attempts += 1
-            time.sleep(1)
-        else:
-            mc.delete("foo")
-            mc.disconnect_all()
-            break
-
-    yield
-
-    subprocess.check_call(["docker-compose", "-f", docker_compose_file, "down"])
-
-
-@pytest.fixture(scope="session")
-def memcached_socket():
-    """
-    Start a standalone Memcached server.
-    """
-    env = os.environ
-    env['PWD'] = HERE
-    env['DOCKER_SOCKET_DIR'] = DOCKER_SOCKET_DIR
-    env['DOCKER_SOCKET_PATH'] = DOCKER_SOCKET_PATH
-    env['UNIXSOCKET_DIR'] = UNIXSOCKET_DIR
-
-    os.chmod(UNIXSOCKET_DIR, 00777)
-
-    docker_compose_file = os.path.join(HERE, 'compose', 'docker-compose.yaml')
-    subprocess.check_call(["docker-compose", "-f", docker_compose_file, "up", "-d", "memcached_socket"], env=env)
-
-    attempts = 0
-    while True:
-        if attempts > 10:
-            raise Exception("Memcached boot timed out!")
-
-        mc = bmemcached.Client(UNIXSOCKET_PATH, USERNAME, PASSWORD)
-        try:
-            mc.set("foo", "bar")
-        except MemcachedException:
-            attempts += 1
-            time.sleep(1)
-        else:
-            mc.delete("foo")
-            mc.disconnect_all()
-            break
-
-    yield
-
-    subprocess.check_call(["docker-compose", "-f", docker_compose_file, "down"])
-
-
-@pytest.fixture
-def client():
-    return bmemcached.Client(["{}:{}".format(HOST, PORT)], USERNAME, PASSWORD)
-
-
-@pytest.fixture
-def client_socket():
-    return bmemcached.Client(UNIXSOCKET_PATH, USERNAME, PASSWORD)
-
-
-@pytest.fixture
-def check():
-    return Memcache('mcache', None, {}, [{}])
-
-
-@pytest.fixture
-def instance():
-    return {
-        'url': "{}".format(HOST),
-        'port': PORT,
-        'tags': ["foo:bar"],
-        'username': USERNAME,
-        'password': PASSWORD,
-    }
-
-
-@pytest.fixture
-def instance_socket():
-    return {
-        'socket': UNIXSOCKET_PATH,
-        'tags': ["foo:bar"],
-        'username': USERNAME,
-        'password': PASSWORD,
-    }
-
-
-@pytest.fixture
-def aggregator():
-    from datadog_checks.stubs import aggregator
-    aggregator.reset()
-    return aggregator
+from common import (PORT, SERVICE_CHECK, HOST, HOST_SOCKET_PATH, GAUGES, RATES, ITEMS_RATES, ITEMS_GAUGES, SLABS_RATES,
+                    SLABS_GAUGES, SLABS_AGGREGATES)
 
 
 def count_connections(port):
@@ -277,7 +71,7 @@ def test_service_with_socket_ok(check, instance_socket, aggregator, memcached_so
     """
     Service is up
     """
-    tags = ["host:unix", "port:{}".format(UNIXSOCKET_PATH), "foo:bar"]
+    tags = ["host:unix", "port:{}".format(HOST_SOCKET_PATH), "foo:bar"]
 
     check.check(instance_socket)
     assert len(aggregator.service_checks(SERVICE_CHECK)) == 1
